refactor(ts): Steps
