+++ source/components/Steps/Steps.tsx
@@ -1,0 +1,177 @@
+/* eslint react/no-did-mount-set-state: 0 */
+import * as React from 'react';
+import PropTypes from 'prop-types';
+import { findDOMNode } from 'react-dom';
+import classNames from 'classnames';
+import debounce from 'lodash/debounce';
+import { isFlexSupported } from './utils';
+import Step, { StepProps } from './Step';
+
+type StepsProps = {
+  direction?: string;
+  labelPlacement?;
+  size?: string;
+  current?: number;
+  initial?: number;
+} & StepProps;
+
+interface StepsState {
+  flexSupported: boolean;
+  lastStepOffsetWidth: number;
+}
+
+export default class Steps extends React.Component<StepsProps, StepsState> {
+  static propTypes = {
+    prefixCls: PropTypes.string,
+    className: PropTypes.string,
+    iconPrefix: PropTypes.string,
+    direction: PropTypes.string,
+    labelPlacement: PropTypes.string,
+    children: PropTypes.node,
+    status: PropTypes.string,
+    size: PropTypes.string,
+    progressDot: PropTypes.oneOfType([PropTypes.bool, PropTypes.func]),
+    style: PropTypes.object,
+    initial: PropTypes.number,
+    current: PropTypes.number,
+    icons: PropTypes.shape({
+      finish: PropTypes.node,
+      error: PropTypes.node
+    })
+  };
+  static defaultProps = {
+    prefixCls: 'fishd-steps',
+    iconPrefix: 'fishd',
+    direction: 'horizontal',
+    labelPlacement: 'horizontal',
+    initial: 0,
+    current: 0,
+    status: 'process',
+    size: '',
+    progressDot: false
+  };
+
+  static Step = Step;
+  private calcTimeout;
+
+  constructor(props) {
+    super(props);
+    this.state = {
+      flexSupported: true,
+      lastStepOffsetWidth: 0
+    };
+    this.calcStepOffsetWidth = debounce(this.calcStepOffsetWidth, 150);
+  }
+  componentDidMount() {
+    this.calcStepOffsetWidth();
+    if (!isFlexSupported()) {
+      this.setState({
+        flexSupported: false
+      });
+    }
+  }
+  componentDidUpdate() {
+    this.calcStepOffsetWidth();
+  }
+  componentWillUnmount() {
+    if (this.calcTimeout) {
+      clearTimeout(this.calcTimeout);
+    }
+
+    // @ts-ignore
+    if (this.calcStepOffsetWidth && this.calcStepOffsetWidth.cancel) {
+      // @ts-ignore
+      this.calcStepOffsetWidth.cancel();
+    }
+  }
+  calcStepOffsetWidth = (): any => {
+    if (isFlexSupported()) {
+      return;
+    }
+    // Just for IE9
+    const domNode = findDOMNode(this);
+    if ((domNode as HTMLElement).children.length > 0) {
+      if (this.calcTimeout) {
+        clearTimeout(this.calcTimeout);
+      }
+      this.calcTimeout = setTimeout(() => {
+        // +1 for fit edge bug of digit width, like 35.4px
+        const lastStepOffsetWidth = ((domNode.lastChild as HTMLElement).offsetWidth || 0) + 1;
+        // Reduce shake bug
+        if (
+          this.state.lastStepOffsetWidth === lastStepOffsetWidth ||
+          Math.abs(this.state.lastStepOffsetWidth - lastStepOffsetWidth) <= 3
+        ) {
+          return;
+        }
+        this.setState({ lastStepOffsetWidth });
+      });
+    }
+  };
+  render() {
+    const {
+      prefixCls,
+      style = {},
+      className,
+      children,
+      direction,
+      labelPlacement,
+      iconPrefix,
+      status,
+      size,
+      current,
+      progressDot,
+      initial,
+      icons,
+      ...restProps
+    } = this.props;
+    const { lastStepOffsetWidth, flexSupported } = this.state;
+    const filteredChildren = React.Children.toArray(children).filter(c => !!c);
+    const lastIndex = filteredChildren.length - 1;
+    const adjustedlabelPlacement = progressDot ? 'vertical' : labelPlacement;
+    const classString = classNames(prefixCls, `${prefixCls}-${direction}`, className, {
+      [`${prefixCls}-${size}`]: size,
+      [`${prefixCls}-label-${adjustedlabelPlacement}`]: direction === 'horizontal',
+      [`${prefixCls}-dot`]: !!progressDot
+    });
+
+    return (
+      <div className={classString} style={style}>
+        {React.Children.map(filteredChildren, (child: React.ReactElement<StepsProps>, index) => {
+          if (!child) {
+            return null;
+          }
+          const stepNumber = initial + index;
+          const childProps = {
+            stepNumber: `${stepNumber + 1}`,
+            prefixCls,
+            iconPrefix,
+            wrapperStyle: style,
+            progressDot,
+            icons,
+            ...child.props
+          };
+          if (!flexSupported && direction !== 'vertical' && index !== lastIndex) {
+            childProps.itemWidth = `${100 / lastIndex}%`;
+            childProps.adjustMarginRight = -Math.round(lastStepOffsetWidth / lastIndex + 1);
+          }
+          // fix tail color
+          if (status === 'error' && index === current - 1) {
+            childProps.className = `${prefixCls}-next-error`;
+          }
+          if (!child.props.status) {
+            if (stepNumber === current) {
+              childProps.status = status;
+            } else if (stepNumber < current) {
+              childProps.status = 'finish';
+            } else {
+              childProps.status = 'wait';
+            }
+          }
+          // @ts-ignore
+          return React.cloneElement(child, childProps);
+        })}
+      </div>
+    );
+  }
+}
