+++ source/components/Steps/index.tsx
@@ -2,7 +2,5 @@ import Steps from './Steps';
 import Step from './Step';
 import './style/index.less';
 
-Steps.Step = Step;
-
 export { Step };
 export default Steps;
